Fixes + Multiple devices
Fix hardcoding of device left from debugging.

Add ability to connect to multiple devices. Each device will be
accessible by a path corresponding to its id.
i.e. NodOSVRPlugin/0/interface, for the first device
NodOSVRPlugin/1/interface, for the second

Fix OpenSpatial Dll to send the id of the device, and use that id
to update the appropriate state.